Fix code review issues: correct constants, add null checks, track emergencies, fix docs

+++ src/main/java/com/adastrea/assistant/GameStateIntegrationDemo.java
@@ -107,7 +107,7 @@ public class GameStateIntegrationDemo {
     
     private static void sleep(int seconds) {
         try {
-            Thread.sleep(seconds * 1000);
+            Thread.sleep((long) seconds * 1000L);
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
         }
